--- add_document/pdf_handler.py
@@ -1,2 +1,0 @@
-def process_pdf(pdf_data,scope):
-    pass
